# [mypyc] Tokenizer for printf-style format string (#10808)