We add a special path for %-style string formatting, which previously is compiled by calling a % binary op. 

The `translate_str_format_percent_sign` serves as an interface from `transform_op_expr` to a specializer for string formatting. A regex tokenizer for %-style format string would parse the original strings into two lists, one is literals and the other is formatting operations.

Currently we directly match the `List[str]` with conversions only for %s and %d cases. In the following PRs, `FormatOp` would be used for better code structure.

`join_formatted_strings`, shared with `str.format()`, merges two lists into one final string using `str_build_op` which is much faster than `str.join`.

## Test Plan

This PR adds several run tests and irbuilding tests. 

The performance on microbenchmarks is: 

master branch:
```
running str_format_percent_operator
..........
interpreted: 0.407026s (avg of 5 iterations; stdev 1.9%)
compiled:    0.309361s (avg of 5 iterations; stdev 2.2%)

compiled is 1.316x faster
```

this PR:
```
running str_format_percent_operator
..........
interpreted: 0.385362s (avg of 5 iterations; stdev 0.75%)
compiled:    0.139198s (avg of 5 iterations; stdev 0.39%)

compiled is 2.768x faster
```

### Have you read the [Contributing Guidelines](https://github.com/python/mypy/blob/master/CONTRIBUTING.md)?

(Once you have, delete this section. If you leave it in, your PR may be closed without action.)

### Description

<!--
If this pull request closes or fixes an issue, write Closes #NNN" or "Fixes #NNN" in that exact
format.
-->

(Explain how this PR changes mypy.)

## Test Plan

<!--
If this is a documentation change, rebuild the docs (link to instructions) and review the changed pages for markup errors.
If this is a code change, include new tests (link to the testing docs). Be sure to run the tests locally and fix any errors before submitting the PR (more instructions).
If this change cannot be tested by the CI, please explain how to verify it manually.
-->

(Write your test plan here. If you changed any code, please provide us with clear instructions on how you verified your changes work.)